% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dct_rate.R
\name{expand_dct_rate}
\alias{expand_dct_rate}
\title{Get the rate of change (first derivative) of DCT}
\usage{
expand_dct_rate(coefs, size = length(coefs))
}
\arguments{
\item{coefs}{DCT coefficients}

\item{size}{length of the result}
}
\value{
A vector of length \code{size} of the rate of change
}
\description{
This will return the rate of change (first derivative)
of a Discrete Cosine Transform. For large \code{size}s,
the first-forward-difference of the inverse DCT will
approximate the results of this function.
}
\examples{
library(faveDCT)
params = c(323.25, 33.41, 2.33, 3.51, -5.10)

# the DCT expansion
expanded = expand_dct(params, size = 1000)

# the first forward difference
ffd = diff(expanded)

# the rate of change
rate = expand_dct_rate(params, size = 1000)

plot(rate, lwd = 5, type = 'l')
lines(ffd, col = "red", lwd = 2)
legend(
  x = 0,
  y = -0.25,
  legend = c("rate", "ffd"),
  col = c("black", "red"),
  lty = 1
)
}
