% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{formant_params}
\alias{formant_params}
\title{Formant DCT paramers}
\format{
An object of class \code{data.frame} with 1365 rows and 9 columns.
}
\usage{
formant_params
}
\description{
Formant DCT paramers
}
\keyword{datasets}
