% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_curve.R
\name{expand_dct}
\alias{expand_dct}
\title{Expand DCT}
\usage{
expand_dct(coefs, size = length(coefs))
}
\arguments{
\item{coefs}{Either a 1d vector of DCT coefficients,
or a \eqn{n\times}\code{coefs} matrix}

\item{size}{The desired output length of the expansion}
}
\value{
A \eqn{n\times}\code{size} matrix of the
         expanded DCT
}
\description{
Given a set of DCT parameters, expand
the fitted curve.
}
\examples{
library(faveDCT)
params = c(323.25, 33.41, 2.33, 3.51, -5.10)

one_d <- expand_dct(params, size = 100)
plot(t(one_d), type = 'l')
}
