% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dct_rate.R
\name{expand_dct_rate}
\alias{expand_dct_rate}
\title{Get the rate of change (first derivative) of DCT}
\usage{
expand_dct_rate(coefs, size = length(coefs))
}
\description{
This will return the rate of change (first derivative)
of a Discrete Cosine Transform. For large \code{size}s,
the first-forward-difference of the inverse DCT will
approximate the results of this function.
}
